moved checkbox classes to UI.helper
