+++ TLM/TLM/TLM.csproj
@@ -228,6 +228,9 @@
     <Compile Include="Traffic\ExtVehicleType.cs" />
     <Compile Include="Traffic\Impl\SegmentEnd.cs" />
     <Compile Include="UI\CustomKeyHandler.cs" />
+    <Compile Include="UI\Helpers\CheckboxOption.cs" />
+    <Compile Include="UI\Helpers\SerializableOptionBase.cs" />
+    <Compile Include="UI\Helpers\SerializableUIOptionBase.cs" />
     <Compile Include="UI\IncompatibleModsPanel.cs" />
     <Compile Include="UI\RemoveCitizenInstanceButtonExtender.cs" />
     <Compile Include="UI\RemoveVehicleButtonExtender.cs" />
