+++ TLM/TLM/UI/Helpers/SerializableOptionBase.cs
@@ -1,0 +1,6 @@
+namespace TrafficManager.UI.Helpers {
+    public abstract class SerializableOptionBase {
+        public abstract void Load(byte data);
+        public abstract byte Save();
+    }
+}
